Version 2.1.1 - Vollstaendige Pruefung auf PHP 8.3 und Contao 4.13/5.7
Ergebnis einer Pruefung des gesamten Bundles mit PHP 8.3 (Syntax) und PHPStan
gegen Contao 4.13.58 und Contao 5.7.7. Gefunden wurden mehrere Stellen, die
unter Contao 5 zur Laufzeit abgestuerzt waeren, die aber Daten voraussetzen und
deshalb von den bisherigen Tests nicht beruehrt wurden.

* Fix: Bewerbungszaehler in der Turnierliste benutzte Controller::getImage() und generateImage() - beide in Contao 5 entfallen. Die Turnierliste war dort nicht mehr aufrufbar, sobald ein Turnier Bewerbungen hatte
* Fix: Brotkrumennavigation der Turniere und Konten benutzte die entfallene Konstante TL_FILES_URL
* Fix: E-Mail-Sendeformular benutzte die entfallene Konstante TL_SCRIPT
* Fix: $this->Database und $this->Environment waren in mehreren DCA-Dateien nie ueber import() geladen; in Contao 5 liefert der Zugriff null
* Fix: Helper::getMeldungen() wurde statisch aufgerufen, war aber nicht static - unter PHP 8 ein Fatal Error in der Teilnehmerliste. Ein Semikolon hinter der if-Bedingung setzte zudem den Zwischenspeicher ausser Kraft
* Fix: Helper::checkResetbuchungen() fragte mit der nicht vorhandenen Variablen $id ab
* Fix: Spezialfilter im Serienmail-Verteiler war durch eine Zuweisungskette wirkungslos
* Fix: Turnierimport suchte nach $set['titel'] statt $set['title'] und legte deshalb Doppelgaenger an
* Fix: ICCF-Import brach ab, wenn die hochgeladene Datei keine CSV-Datei war
* Fix: tl_settings enthielt den Schluessel mandatory zweimal
* Fix: Statistik-Laufzeitmessung rechnete gegen eine undefinierte Variable
* Fix: Parameter differenz der Turnierstatistik wird vor der Rechnung in eine Zahl gewandelt
* Change: Ueberfluessiger zweiter Parameter am Konstruktor von Contao\File entfernt
* Change: Ungenutzter Dienst request_stack aus dem ICCF-Import-Controller entfernt
* Change: Fehlerhafte @param- und @return-Angaben richtiggestellt

+++ src/Classes/Turnierstatistik.php
@@ -55,7 +55,9 @@ class Turnierstatistik
 		$urlJahr = (int)Input::get('jahr');
 		$urlMonat = (int)Input::get('monat');
 		$urlTag = (int)Input::get('tag');
-		$differenz = Input::get('differenz');
+		// Der Wert wandert in Rechnungen; ohne die Umwandlung könnte ein
+		// Aufruf wie ?differenz[]=1 unter PHP 8 einen TypeError auslösen.
+		$differenz = (int) Input::get('differenz');
 
 		// Neues Datum für anzuzeigende Daten ermitteln
 		if(!$urlJahr && !$urlMonat && !$urlTag)
